split out a DiagnosedThankYou view to be displayed conditionally
